+++ packages/google/src/google-generative-ai-settings.ts
@@ -22,6 +22,7 @@ export type GoogleGenerativeAIModelId =
   // Experimental models
   // https://ai.google.dev/gemini-api/docs/models/experimental-models
   | 'gemini-2.5-pro-exp-03-25'
+  | 'gemini-2.5-pro-preview-05-06'
   | 'gemini-2.5-flash-preview-04-17'
   | 'gemini-exp-1206'
   | 'gemma-3-27b-it'