+++ packages/elevenlabs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/elevenlabs
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+
 ## 0.0.2
 
 ### Patch Changes