+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/perplexity
 
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+
 ## 1.1.8
 
 ### Patch Changes