+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/valibot
 
+## 0.1.27
+
+### Patch Changes
+
+- ai@4.3.15
+
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies [a295521]
+  - ai@4.3.14
+
 ## 0.1.25
 
 ### Patch Changes